refactor(livechat): reduce ChatContainer to a lifecycle-only shell
The connector now renders the presentational component directly and keeps
ChatContainer purely for its class lifecycle (render returns null),
rendering both as siblings. Trim ChatContainerProps to just what the
lifecycle needs, dropping the handler props it no longer uses.
